# backend/.vscode/ZA_swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "All API",
    "description": "TODO: Add Description",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://example.com",
      "variables": {}
    }
  ],
  "paths": {
    "/api/login": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "GetAuthorizationToken",
        "description": "TODO: Add Description",
        "operationId": "GetAuthorizationToken",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAuthorizationTokenRequest"
              },
              "example": {
                "username": "קוד_אואזיס_0507406078",
                "password": "0507406078"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/OldCustomer/getTicketHistory": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "GetTicketHistory",
        "description": "TODO: Add Description",
        "operationId": "GetTicketHistory",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 35
            }
          },
          {
            "name": "ticketSystem",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "RoadProtectIL"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/OldTicket/addTicket": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "AddTicket",
        "description": "TODO: Add Description",
        "operationId": "AddTicket",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTicketRequest"
              },
              "example": {
                "citationNo": "98798911935",
                "country": "3",
                "licensePlate": "4819033",
                "courthouse": "10",
                "violationDate": "2019-08-11 18:04:43",
                "violationCodes": "11043",
                "institutionId": "1",
                "modelCode": "Ceed",
                "makeCode": "Kia",
                "modelType": "Ceed",
                "vehicleMake": "Kia",
                "vehicleYear": "2011",
                "ticketStatus": "34",
                "city": "תל אביב - יפו",
                "address": "שדרות רוטשילד 98",
                "violationId": "10001",
                "customerId": "71",
                "ticketSystem": "RoadProtectIL",
                "amount": "250",
                "currency": "ILS",
                "questionsAndAnswers": [
                  [
                    "topicId-001",
                    "reasonId-001",
                    "7"
                  ],
                  [
                    "topicId-001",
                    "reasonId-002",
                    "reasonId-002-043",
                    "answerId-002-043-1",
                    "16"
                  ],
                  [
                    "topicId-001",
                    "reasonId-003",
                    "reasonId-003-043",
                    "answerId-003-043-1",
                    "16"
                  ]
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/Ticket/updateTicket": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "updateTicket",
        "description": "TODO: Add Description",
        "operationId": "UpdateTicket",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateTicketRequest"
              },
              "example": {
                "ticketType": null,
                "citationNo": "987987",
                "country": "3",
                "licensePlate": "4819033",
                "courthouse": null,
                "violationDate": "2019-08-11 18:04:43",
                "violationCodes": null,
                "institutionId": "1",
                "modelCode": "Ceed",
                "makeCode": "Kia",
                "modelType": "Ceed",
                "vehicleMake": "Kia",
                "vehicleYear": "2011",
                "ticketStatus": 40,
                "city": "תל אביב - יפו",
                "address": "שדרות רוטשילד 98",
                "violationId": "10001",
                "customerId": "35",
                "ticketSystem": "RoadProtectIL",
                "transactionState": 40,
                "amount": "280",
                "currency": "ILS",
                "id": "455",
                "hasMembership": false,
                "finished": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/Ticket/getTicketDocumentTypeList": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "getTicketDocumentTypeList",
        "description": "TODO: Add Description",
        "operationId": "GetTicketDocumentTypeList",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 0
            }
          },
          {
            "name": "totalInstitutes",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/Ticket/uploadTicketDocument": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "uploadTicketDocument",
        "description": "TODO: Add Description",
        "operationId": "UploadTicketDocument",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/uploadTicketDocumentRequest"
              },
              "example": {
                "ticketId": 42,
                "ticketDocumentBase64": "iVBORw0KGgoAAAANSUhEUgAAAOEAAADhCAMAAAAJbSJIAAAAb1BMVEX///8AAABQUFCioqJKSkp5eXmdnZ0oKCjr6+uKioq8vLze3t5bW1vn5+ezs7PT09OUlJQ5OTkuLi4eHh54eHg2Nja+vr5paWlubm7Ly8vR0dEVFRW9vb1AQECZmZmGhoYLCwv29varq6sSEhJhYWHHkF97AAADLUlEQVR4nO3d7VbaQBRGYcKHBAIBBFG0VlHv/xoLyKoiYXVOOe8cavfz27OYLclM4h9bLQAAAAC4KGU9aM/657vp9uoqOuZYOZ0XrnqXFVktfPN23ibRWR+mgr6txTq67N24LwrcuI+O21rp+jbq6LxWaykNLIqX6MB7cWD4t1jKA4sidkt9zFD4EBl4lbTEblU2eklMfIoLTLxGT275s8TEuMebXtoCT+4VD4mFo5xRn6VuM6+9QaN24nzcZqN6WDt2F1SYepE5iHlAneQLDDr2812kmzepkMLrjIVFSGHOwJAjMedtWBTLgEL1a9OhiJeoOmthxLNp6nOzj9uAws7+s3ta+z9StgMLxR/T/vaFXQp1KHRCoRCFTigUotAJhUIUOqFQiEInFApR6IRCIQqdUChEoRMKhSh0QqEQhU4oFKLQCYVCFDqhUIhCJxQKUeiEQiEKnVAoRKETCoUodEKhEIVOKBSi0AmFQhQ6oVCIQicUClHohEIhCp1QKEShEwqFKHRCoRCFTigUotAJhUIUOqFQiEInFApR6IRCIQqdUChEoRMKhSh0QqEQhU4oFKLQCYVCFDqhUIhCJxQKUeiEQiEKnVAoRKETCoUodEKhEIVOKBSi0AmFQhQ6oVCIQicUCmUqvID/6dy5UupcxxfmQSGFFFKYqj+o63o6+7aFo2o/Or79noUvn4aH/0JhbVzj8GB6ZZx+u/zCznnjdwGFS9MKj6+ykWl+GlBYmVa4Opq/N83XAYUtywL7DfPX5/2GMrAca7cN83eWwux1W1PDAn82zFuOm8fsdVsTwwq/7qRbPwzzV9nrdh7SV9i0FVpO/XH2uB3DZdp0Hw7Sx0fZ296N05c4axhvp48vs7ftLdLXWB0Nl+nDMfuMcZGLo2HDNR72FZpWOfkyangkiroLd26SlzkvDwbX/fTCdVDcjmGzOUgsDY9sw5OfnoXlDWP4V1NBh/0Hy7n92KnG5bjqdA0zEa9NZyRuPNt+vOmBNruJcdEWwffgb4aT36Rd/vmzM1lZ7qxU/ZD3+pNWpvfZBKNh6DHYaDloz18d2p7n3af68vIAAAAA/N9+ASPSNE8ENa/ZAAAAAElFTkSuQmCC",
                "ticketDocumentMimeType": "image/png",
                "documentDescription": "Test Description",
                "documentTypeCode": 1060
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/oldCustomer/updateCustomerProfile": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "updateCustomerProfile",
        "description": "TODO: Add Description",
        "operationId": "UpdateCustomerProfile",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateCustomerProfileRequest"
              },
              "example": {
                "firstName": "דערוק",
                "lastName": "מעיער",
                "mobileNumber": "(111)111-5566",
                "email": "sample@email.com",
                "address": "sampleAddress",
                "city": "sampleCity",
                "customerId": "33",
                "israelIdNumber": "1234567891235"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/Ticket/getTicketStatusList": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "getTicketStatusList",
        "description": "TODO: Add Description",
        "operationId": "GetTicketStatusList",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 0
            }
          },
          {
            "name": "totalInstitutes",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "GetAuthorizationTokenRequest": {
        "title": "GetAuthorizationTokenRequest",
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "username": "קוד_אואזיס_0507406078",
          "password": "0507406078"
        }
      },
      "AddTicketRequest": {
        "title": "AddTicketRequest",
        "required": [
          "citationNo",
          "country",
          "licensePlate",
          "courthouse",
          "violationDate",
          "violationCodes",
          "institutionId",
          "modelCode",
          "makeCode",
          "modelType",
          "vehicleMake",
          "vehicleYear",
          "ticketStatus",
          "city",
          "address",
          "violationId",
          "customerId",
          "ticketSystem",
          "amount",
          "currency",
          "questionsAndAnswers"
        ],
        "type": "object",
        "properties": {
          "citationNo": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "licensePlate": {
            "type": "string"
          },
          "courthouse": {
            "type": "string"
          },
          "violationDate": {
            "type": "string"
          },
          "violationCodes": {
            "type": "string"
          },
          "institutionId": {
            "type": "string"
          },
          "modelCode": {
            "type": "string"
          },
          "makeCode": {
            "type": "string"
          },
          "modelType": {
            "type": "string"
          },
          "vehicleMake": {
            "type": "string"
          },
          "vehicleYear": {
            "type": "string"
          },
          "ticketStatus": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "violationId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "ticketSystem": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "questionsAndAnswers": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "example": {
          "citationNo": "98798911935",
          "country": "3",
          "licensePlate": "4819033",
          "courthouse": "10",
          "violationDate": "2019-08-11 18:04:43",
          "violationCodes": "11043",
          "institutionId": "1",
          "modelCode": "Ceed",
          "makeCode": "Kia",
          "modelType": "Ceed",
          "vehicleMake": "Kia",
          "vehicleYear": "2011",
          "ticketStatus": "34",
          "city": "תל אביב - יפו",
          "address": "שדרות רוטשילד 98",
          "violationId": "10001",
          "customerId": "71",
          "ticketSystem": "RoadProtectIL",
          "amount": "250",
          "currency": "ILS",
          "questionsAndAnswers": [
            [
              "topicId-001",
              "reasonId-001",
              "7"
            ],
            [
              "topicId-001",
              "reasonId-002",
              "reasonId-002-043",
              "answerId-002-043-1",
              "16"
            ],
            [
              "topicId-001",
              "reasonId-003",
              "reasonId-003-043",
              "answerId-003-043-1",
              "16"
            ]
          ]
        }
      },
...
      "updateCustomerProfileRequest": {
        "title": "updateCustomerProfileRequest",
        "required": [
          "firstName",
          "lastName",
          "mobileNumber",
          "email",
          "address",
          "city",
          "customerId",
          "israelIdNumber"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "mobileNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "israelIdNumber": {
            "type": "string"
          }
        },
        "example": {
          "firstName": "דערוק",
          "lastName": "מעיער",
          "mobileNumber": "(111)111-5566",
          "email": "sample@email.com",
          "address": "sampleAddress",
          "city": "sampleCity",
          "customerId": "33",
          "israelIdNumber": "1234567891235"
        }
      },
...
    }
  },
...
}